+++ packages/spotlight/CHANGELOG.md
@@ -10,6 +10,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ## [1.10.0] - 2017-10-09
 
+- `spotlight` to update pointer mode during the event capture phase for pointer hide and show events
+
 ### Fixed
 
 - `spotlight.Spotlight` method `focus()` to verify that the target element matches its container's selector rules prior to setting focus